requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.4
Django==4.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.4.4
# mysqlclient==2.2.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
